+++ lib/contur.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+require_relative 'contur/version'
+
+# require application files
+require_relative 'contur/bindable_hash'
+require_relative 'contur/constants'
+require_relative 'contur/controller'
+require_relative 'contur/config'