+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This CHANGELOG follows [Keep a CHANGELOG](http://keepachangelog.com/).
 
+## [0.0.3] - 2016-10-24
+### Added
+- Travis support
+- Accept config file with both `.yml` and `.yaml` extensions
+
+### Changed
+- Project name changed to Contur
+
 ## [0.0.2] - 2016-10-20
 ### First open-source release